Rename main
